+++ src/game/options.ts
@@ -1,3 +1,4 @@
 export const Options = {
     HighFPS: true,
+    SmoothCamera: true,
 }